Share container listings between sidebar tree and main browse
