+++ app/lib/hooks/useBrowseStorage.ts
@@ -2,11 +2,9 @@
 
 import { useEffect, useState } from "react";
 import { getAuthenticatedSession } from "../helpers";
-import { getSolidDataset, toRdfJsDataset } from "@inrupt/solid-client";
 import { FileItemData } from "../../components/FileItem";
-import { ContainerDataset } from "../class/ContainerDataset";
-import { DataFactory } from "n3";
 import { loadContainerListing } from "../cache";
+import { fetchContainerListing } from "../helpers/containerListingUtils";
 
 interface UseBrowseStorageResult {
   files: FileItemData[];
@@ -14,69 +12,6 @@ interface UseBrowseStorageResult {
   error: Error | null;
 }
 
-async function fetchContainerListing(
-  url: string,
-  fetchFn: typeof fetch,
-): Promise<FileItemData[]> {
-  const container = new ContainerDataset(
-    toRdfJsDataset(await getSolidDataset(url, { fetch: fetchFn })),
-    DataFactory,
-  ).container;
-
-  if (container === undefined) {
-    throw new Error("Container not found");
-  }
-
-  const fileItems: FileItemData[] = [];
-
-  for (const item of container.contains) {
-    try {
-      let mimeType = item.mimeType;
-
-      // For non-folder filed, only fetch content-type if not already in the RDF metadata
-      if (!item.isContainer && !mimeType) {
-        try {
-          const headResponse = await fetchFn(item.id, {
-            method: "HEAD",
-            headers: {
-              Accept: "*/*",
-            },
-          });
-
-          if (headResponse.ok) {
-            const contentType = headResponse.headers.get("Content-Type");
-            if (contentType) {
-              mimeType = contentType.split(";")[0].trim();
-            }
-          }
-        } catch (err) {
-          console.debug(`Could not fetch content-type for ${item}:`, err);
-        }
-      }
-
-      fileItems.push({
-        id: item.id,
-        name: item.name,
-        type: item.fileType,
-        url: item.id,
-        lastModified: item.lastModified,
-        size: item.size,
-        mimeType,
-      });
-    } catch (err) {
-      console.error(`Failed to process item ${item}:`, err);
-    }
-  }
-
-  // sort by folder first then in alphabetical order using the name
-  fileItems.sort((a, b) => {
-    if (a.type === "folder" && b.type !== "folder") return -1;
-    if (a.type !== "folder" && b.type === "folder") return 1;
-    return a.name.localeCompare(b.name);
-  });
-
-  return fileItems;
-}
 
 /**
  * Hook to browse/list the contents of a Solid storage container
